# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hospital Drug Inventory Management System",
    "description": "AI-powered drug inventory optimization using demand prediction and clustering",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "description": "Authenticate user and return JWT token.\nWorks for both hospital and vendor users.\n\n**Request Body:**\n- `user_email`: User email address\n- `user_password`: User password\n\n**Returns:**\n- `access_token`: JWT token for authenticated requests\n- `token_type`: Bearer token type\n- `user_id`: User identifier\n- `user_name`: User name\n- `organization_id`: Hospital/Vendor/Organization ID\n- `user_role`: User role (hospital_admin, hospital_staff, vendor_admin, vendor_staff)",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Current User Info",
        "description": "Get current logged-in user details.\nWorks for both hospital and vendor users.\n\n**Headers:**\n- `Authorization: Bearer \u003Ctoken\u003E`\n\n**Returns:**\n- User details with organization information",
        "operationId": "get_current_user_info_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/hospital/calculate-predictions": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Calculate Predictions",
        "description": "Calculate X1, X2, X3, X4 and inventory parameters for all medicines (Hospital users only).\n\n**Authentication:** Requires hospital_admin or hospital_staff role\n\n**Implementation Steps:**\n1. Fetch stock and usage data from database\n2. Calculate X1_AMC (Average Monthly Consumption)\n3. Calculate X2_Prescriptions (Monthly average prescriptions)\n4. Calculate X3_CDPR (Chronic Disease Prescription Ratio)\n5. Calculate X4_CV (Coefficient of Variation)\n6. Apply K-Means clustering (4 clusters)\n7. Calculate (s, S) inventory parameters\n8. Save predictions to database",
        "operationId": "calculate_predictions_api_hospital_calculate_predictions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/hospital/predictions": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Predictions",
        "description": "Retrieve calculated predictions with filters (Hospital users only).\n\n**Authentication:** Requires hospital_admin or hospital_staff role\n\n**Query Parameters:**\n- `cluster_group`: Filter by cluster (1, 2, 3, or 4)\n- `low_stock_only`: Boolean - show only items below reorder point\n- `page`: Page number (default: 1)\n- `limit`: Items per page (default: 50)",
        "operationId": "get_predictions_api_hospital_predictions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Filter by cluster group (1-4)",
              "title": "Cluster Group"
            },
            "description": "Filter by cluster group (1-4)"
          },
          {
            "name": "low_stock_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Show only items below reorder point",
              "default": false,
              "title": "Low Stock Only"
            },
            "description": "Show only items below reorder point"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hospital/alerts": {
      "get": {
        "tags": [
          "alerts and orders"
        ],
        "summary": "Get Alerts",
        "description": "Get system alerts with filters\n\n**Query Parameters:**\n- `status`: unread, read, resolved\n- `severity`: critical, high, medium, low\n- `alert_type`: low_stock, expiry_warning, order_delayed\n- `limit`: Maximum number of alerts (default: 50)",
        "operationId": "get_alerts_api_hospital_alerts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter: unread, read, resolved",
              "title": "Status"
            },
            "description": "Filter: unread, read, resolved"
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter: critical, high, medium, low",
              "title": "Severity"
            },
            "description": "Filter: critical, high, medium, low"
          },
          {
            "name": "alert_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter: low_stock, expiry_warning, order_delayed",
              "title": "Alert Type"
            },
            "description": "Filter: low_stock, expiry_warning, order_delayed"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/hospital/orders": {
      "get": {
        "tags": [
          "alerts and orders"
        ],
        "summary": "Get Orders",
        "description": "List all orders with filters\n\n**Query Parameters:**\n- `status`: pending, completed, cancelled\n- `from_date`: Start date in YYYY-MM-DD format\n- `to_date`: End date in YYYY-MM-DD format\n- `page`: Page number (default: 1)\n- `limit`: Items per page (default: 20)",
        "operationId": "get_orders_api_hospital_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filter: pending, completed, cancelled",
              "title": "Status"
            },
            "description": "Filter: pending, completed, cancelled"
          },
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Start date (YYYY-MM-DD)",
              "title": "From Date"
            },
            "description": "Start date (YYYY-MM-DD)"
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "End date (YYYY-MM-DD)",
              "title": "To Date"
            },
            "description": "End date (YYYY-MM-DD)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertResponse": {
        "properties": {
          "alert_id": {
            "type": "integer",
            "title": "Alert Id"
          },
          "medicine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicine Id"
          },
          "medicine_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicine Name"
          },
          "alert_type": {
            "type": "string",
            "title": "Alert Type"
          },
          "alert_message": {
            "type": "string",
            "title": "Alert Message"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "alert_status": {
            "type": "string",
            "title": "Alert Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "alert_id",
          "alert_type",
          "alert_message",
          "severity",
          "alert_status",
          "created_at"
        ],
        "title": "AlertResponse"
      },
      "AlertsListResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "type": "array",
            "title": "Alerts"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "alerts"
        ],
        "title": "AlertsListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedicinePredictionResponse": {
        "properties": {
          "medicine_id": {
            "type": "string",
            "title": "Medicine Id"
          },
          "medicine_name": {
            "type": "string",
            "title": "Medicine Name"
          },
          "current_stock": {
            "type": "integer",
            "title": "Current Stock"
          },
          "X1_amc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X1 Amc"
          },
          "X2_prescriptions": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X2 Prescriptions"
          },
          "X3_CDPR": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X3 Cdpr"
          },
          "X4_CV": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "X4 Cv"
          },
          "safety_stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Stock"
          },
          "reorder_stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reorder Stock"
          },
          "max_stock": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Stock"
          },
          "cluster_group": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Group"
          },
          "needs_reorder": {
            "type": "boolean",
            "title": "Needs Reorder",
            "default": false
          },
          "suggested_order_quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Order Quantity"
          }
        },
        "type": "object",
        "required": [
          "medicine_id",
          "medicine_name",
          "current_stock"
        ],
        "title": "MedicinePredictionResponse"
      },
...
      "PredictionRequest": {
        "properties": {
          "medicine_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicine Ids"
          },
          "recalculate_all": {
            "type": "boolean",
            "title": "Recalculate All",
            "default": false
          }
        },
        "type": "object",
        "title": "PredictionRequest"
      },
      "PredictionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "medicines_analyzed": {
            "type": "integer",
            "title": "Medicines Analyzed"
          },
          "clusters_formed": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Clusters Formed"
          },
          "calculation_time_seconds": {
            "type": "number",
            "title": "Calculation Time Seconds"
          }
        },
        "type": "object",
        "required": [
          "message",
          "medicines_analyzed",
          "clusters_formed",
          "calculation_time_seconds"
        ],
        "title": "PredictionResponse"
      },
      "PredictionsListResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "medicines": {
            "items": {
              "$ref": "#/components/schemas/MedicinePredictionResponse"
            },
            "type": "array",
            "title": "Medicines"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "page",
          "limit",
          "medicines"
        ],
        "title": "PredictionsListResponse"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "user_role": {
            "type": "string",
            "title": "User Role"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "user_id",
          "user_name",
          "organization_id",
          "user_role"
        ],
        "title": "TokenResponse"
      },
      "UserLogin": {
        "properties": {
          "user_email": {
            "type": "string",
            "format": "email",
            "title": "User Email"
          },
          "user_password": {
            "type": "string",
            "title": "User Password"
          }
        },
        "type": "object",
        "required": [
          "user_email",
          "user_password"
        ],
        "title": "UserLogin"
      },
      "UserResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "organization_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Name"
          },
          "user_role": {
            "type": "string",
            "title": "User Role"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "user_name",
          "user_email",
          "organization_id",
          "user_role"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}